update soft delete for Posts

+++ app/Http/Controllers/API/PostController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\Post\UpdateRequest;
 use App\Http\Traits\ImageStorageTrait;
 use App\Http\Traits\ResponseTrait;
 use App\Models\Post;
+use Illuminate\Http\Request;
 
 class PostController extends Controller
 {
@@ -18,30 +19,47 @@ class PostController extends Controller
     {
         $userId = auth()->user()->id;
         $path = null;
-        if($request->has('image')){
+        if ($request->has('image')) {
             $path = $this->storeImage('posts', $request->file('image'));
         }
 
         $postAttributes = $request->validated();
         $postAttributes['image'] = $path;
         $postAttributes['user_id'] = $userId;
-        Post::query()->create($postAttributes);
 
-        return $this->successResponse(message:  'Tạo bài đăng thành công');
+        $post = Post::query()->create($postAttributes);
+
+        if((int)$postAttributes['status'] !== 0){
+            $data = Post::query()
+                ->where('id', $post->id)
+                ->withCount('likes')
+                ->withCount('dislikes')
+                ->withCount('comments')
+                ->with('user:id,last_name,first_name,avatar')
+                ->with('reactions')
+                ->get();
+
+            return $this->successResponse($data ,'Tạo bài đăng thành công!');
+        }
+
+        return $this->successResponse($post, 'Tạo bài đăng thành công!');
     }
 
     public function edit(Post $post)
     {
-        if($post->user_id === auth()->user()->id){
+        if ($post->user_id === auth()->user()->id) {
             return $this->successResponse($post);
         }
+
         return $this->errorResponse('Bài đăng này không phải của bạn!');
     }
+
     public function update(UpdateRequest $request, $postId)
     {
         $postAttributes = $request->validated();
         $post = Post::query()->find($postId);
-        if ($post->user_id !== auth()->user()->id){
+
+        if ($post->user_id !== auth()->user()->id) {
             return $this->errorResponse('Bài đăng này không phải của bạn!');
         }
 
@@ -54,16 +72,58 @@ class PostController extends Controller
         }
 
         $post->update($postAttributes);
+        if((int)$postAttributes['status'] !== 0){
+            $data = Post::query()
+                ->where('id', $post->id)
+                ->withCount('likes')
+                ->withCount('dislikes')
+                ->withCount('comments')
+                ->with('user:id,last_name,first_name,avatar')
+                ->with('reactions')
+                ->get();
 
-        return $this->successResponse(message:  'Cập nhật bài đăng thành công!');
+            return $this->successResponse($data ,'Cập nhật bài đăng thành công!');
+        }
+
+        return $this->successResponse($post, 'Cập nhật bài đăng thành công!');
+    }
+    public function show(Post $post){
+        $data = Post::query()
+            ->where('id', $post->id)
+            ->withCount('likes')
+            ->withCount('dislikes')
+            ->withCount('comments')
+            ->with('user:id,last_name,first_name,avatar')
+            ->with('reactions')
+            ->get();
+
+        return $this->successResponse($data);
     }
 
     public function destroy(Post $post)
     {
-        $this->deleteImage($post->image);
+        if ($post->user_id === auth()->user()->id) {
+            $post->delete();
 
-        $post->delete();
+            return $this->successResponse(message: 'Xoá thành công!');
+        }
 
-        return $this->successResponse(message: 'Xoá thành công!');
+        return $this->errorResponse('Đây là bài đăng của người khác');
+    }
+
+    public function updatePostStatus(Request $request,$postId){
+        $post = Post::query()->find($postId);
+        if ($post) {
+            $post->update([
+                'status' => $request->status
+            ]);
+            if((int)$request->status === 1){
+                return $this->successResponse(message: 'Hiện bài viết thành công!');
+            }
+
+            return $this->successResponse(message: 'Ẩn bài viết thành công!');
+        }
+
+        return $this->errorResponse('Không tìm thấy bài viết');
     }
 }
